# openapi/opal-osa-integration.yaml
openapi: 3.0.0
info:
  title: OPAL Decision & Diagnostics API
  version: 1.0.0
  description: |
    Complete API specification for OPAL → OSA data flow, decision-making, and troubleshooting.

    This API maintains backward compatibility with existing OPAL integrations while providing
    enhanced decision layer capabilities with evidence-based recommendations, confidence scoring,
    and impact/effort analysis.

    **Key Features:**
    - Idempotent webhook processing for reliable data flow
    - Evidence-based decision making with confidence metrics
    - Comprehensive diagnostics and troubleshooting tools
    - Force sync and orchestration capabilities
    - Backward compatible with existing OPAL workflows
  contact:
    name: OPAL Development Team
    url: https://opal-2025.vercel.app
  license:
    name: MIT
servers:
  - url: http://localhost:3000
    description: Local development server
  - url: https://opal-2025.vercel.app
    description: Production server
paths:
...
  /api/webhooks/opal-workflow:
    post:
      tags:
        - Webhooks
      summary: Idempotent receiver for OPAL agent webhook events
      description: |
        Receives and processes webhook events from OPAL agents with idempotency guarantees.

        **Idempotency:** Duplicate events with the same workflow_id + agent_id + offset
        combination are safely ignored to prevent data corruption.

        **Backward Compatibility:** Maintains existing webhook signature validation
        and processing flows while adding enhanced monitoring and diagnostics.
      operationId: receiveOpalWebhook
      security:
        - hmacSignature: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OpalWebhookEvent'
            examples:
              success_event:
                summary: Successful agent execution
                value:
                  workflow_id: "wf-content-audit-2025"
                  agent_id: "content_review"
                  execution_status: "success"
                  agent_data:
                    findings_count: 47
                    issues_identified: 3
                    confidence_score: 0.89
                  offset: 1
                  timestamp: "2025-11-12T15:30:00.000Z"
              failure_event:
                summary: Failed agent execution
                value:
                  workflow_id: "wf-audience-analysis-2025"
                  agent_id: "audience_suggester"
                  execution_status: "failure"
                  error_details:
                    error_code: "DATA_ACCESS_ERROR"
                    error_message: "Unable to access customer data source"
                  offset: 2
                  timestamp: "2025-11-12T15:35:00.000Z"
      responses:
        '202':
          description: Webhook accepted for processing
          headers:
            X-Correlation-ID:
              schema:
                type: string
              description: Request correlation ID for tracing
            X-Idempotency-Key:
              schema:
                type: string
              description: Idempotency key for duplicate detection
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WebhookResponse'
              example:
                success: true
                message: "Webhook event processed successfully"
                data:
                  event_id: "evt-1699876543210-xyz789"
                  workflow_id: "wf-content-audit-2025"
                  agent_id: "content_review"
                  processing_status: "accepted"
                  duplicate_detected: false
                correlation_id: "api-1699876543210-abc123"
                timestamp: "2025-11-12T15:30:05.000Z"
                duration_ms: 45
        '400':
          description: Invalid webhook payload or signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid or missing webhook signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Duplicate event detected (idempotency conflict)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  schemas:
...
    OpalAgentType:
      type: string
      enum:
        - experiment_blueprinter
        - audience_suggester
        - content_review
        - roadmap_generator
        - integration_health
        - personalization_idea_generator
        - cmp_organizer
        - customer_journey
        - geo_audit
      description: Supported OPAL agent types
    ExecutionStatus:
      type: string
      enum: [success, failure, timeout]
      description: Webhook event execution status
...
    OpalWebhookEvent:
      type: object
      required:
        - workflow_id
        - agent_id
        - execution_status
        - offset
      properties:
        workflow_id:
          type: string
          description: Unique identifier for the workflow
          example: "wf-content-audit-2025"
        agent_id:
          $ref: '#/components/schemas/OpalAgentType'
        execution_status:
          $ref: '#/components/schemas/ExecutionStatus'
        agent_data:
          type: object
          description: Agent-specific execution results and data
          additionalProperties: true
        error_details:
          type: object
          description: Error information (present when execution_status is failure)
          properties:
            error_code:
              type: string
              description: Machine-readable error code
            error_message:
              type: string
              description: Human-readable error description
            stack_trace:
              type: string
              description: Technical stack trace (development only)
        offset:
          type: integer
          minimum: 0
          description: Event sequence number for idempotency and ordering
        timestamp:
          type: string
          format: date-time
          description: Event generation timestamp (ISO 8601)
        retry_count:
          type: integer
          minimum: 0
          description: Number of previous retry attempts
          default: 0
...
    WebhookResponse:
      allOf:
        - $ref: '#/components/schemas/StandardApiResponse'
        - type: object
          properties:
            data:
              type: object
              properties:
                event_id:
                  type: string
                  description: Unique event identifier
                workflow_id:
                  type: string
                  description: Workflow identifier from webhook
                agent_id:
                  $ref: '#/components/schemas/OpalAgentType'
                processing_status:
                  type: string
                  enum: [accepted, rejected, duplicate]
                  description: Webhook processing result
                duplicate_detected:
                  type: boolean
                  description: Whether this was a duplicate event
...
    StandardApiResponse:
      type: object
      required:
        - success
        - correlation_id
        - timestamp
      properties:
        success:
          type: boolean
          description: Whether the request was successful
          example: true
        data:
          description: Response data (varies by endpoint)
        message:
          type: string
          description: Optional human-readable message
        correlation_id:
          type: string
          description: Unique request identifier for tracing
          example: "api-1699876543210-def456"
        timestamp:
          type: string
          format: date-time
          description: Response generation timestamp (ISO 8601 format)
          example: "2025-11-12T15:45:00.000Z"
        duration_ms:
          type: integer
          minimum: 0
          description: Processing time in milliseconds
          example: 245
    ErrorResponse:
      type: object
      required:
        - success
        - error
        - message
        - correlation_id
        - timestamp
      properties:
        success:
          type: boolean
          description: Always false for error responses
          example: false
        error:
          type: string
          description: Error code
          example: "VALIDATION_ERROR"
        message:
          type: string
          description: Human-readable error description
          example: "Invalid request body"
        correlation_id:
          type: string
          description: Unique request identifier for tracing
          example: "api-1699876543210-def456"
        timestamp:
          type: string
          format: date-time
          description: Error occurrence timestamp (ISO 8601 format)
          example: "2025-11-12T15:45:00.000Z"
        duration_ms:
          type: integer
          minimum: 0
          description: Processing time before error occurred
          example: 25
        data:
          description: Additional error details (only in development)
  securitySchemes:
    hmacSignature:
      type: apiKey
      in: header
      name: X-Signature-SHA256
      description: |
        HMAC-SHA256 signature for webhook authentication.

        Format: sha256=<hex_digest>

        The signature is computed using the raw request body and a shared secret key.
  responses:
...
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error: "INTERNAL_SERVER_ERROR"
            message: "An unexpected error occurred"
            correlation_id: "api-1699876543210-def456"
            timestamp: "2025-11-12T15:45:00.000Z"
            duration_ms: 150
...